New Version

+++ VendingMachine/Program.cs
@@ -2,6 +2,8 @@
 
 using System.Diagnostics;
 using VendingMachine;
+using VendingMachine.Drinks;
+using VendingMachine.Ingredients;
 
 class Program
 {
@@ -16,14 +18,14 @@ class Program
     static void TestDrinkCostCalculation()
     {
         Coffee coffeeIngredient = new Coffee("Coffee", 1, 1);
-        Water waterIngredient = new Water("Water", 0.2m, 1);
+        Water waterIngredient = new Water("Water", 0.2, 1);
 
         Espresso<Ingredient> espressoRecipe = new Espresso<Ingredient>(coffeeIngredient, waterIngredient);
 
-        decimal customProfitMargin =1.3m; // Set your custom profit margin here
+        double customProfitMargin =1.3; // Set your custom profit margin here
         Drink<Espresso<Ingredient>, Ingredient> espressoDrink = new Drink<Espresso<Ingredient>, Ingredient>(espressoRecipe, customProfitMargin);
 
-        decimal calculatedCost = Math.Round(espressoDrink.CalculateCost(), 2);
+        double calculatedCost = Math.Round(espressoDrink.CalculateCost(), 2);
 
         // Output the results
         Console.WriteLine($"Calculated Cost: {calculatedCost}");
